+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # ChangeLog for the MySQL Enterprise Backup Wrapper Script 
 
+## 1.1.0 (March 23, 2015)
+
+### Backwards Incompatible Changes
+  - Disabled history logging by default
+
+### Improvements
+  - Added the `--mysqlbackup option` to allow the location of the mysqlbackup binary to be set.
+  - Added the `--history-logging` to turn on history logging.
+  - Backup /etc/my.cnf by default
+
 ## 1.0.0 (December 16, 2014)
 
   - Initial Release